nm-dispatcher: dispatch on method change
Trigger a dispatcher event when the IPv4 or IPv6 method of a connection
has been changed. Some devices such as phones have already a DHCP client
running for accepting connections when they are plugged into USB to
transfer data over SSH. When NetworkManager switches the connection IP
method to shared, it spawns a dnsmasq process to handle DHCP and DNS for
that connection. However, a dispatcher event is needed to disable the
external DHCP server for these USB connections as NetworkManager's
dnsmasq handles them now. Moreover, when the connection method is switched
to a different mode, the external DHCP server needs to be spawned again
to make sure that SSH connections are still possible to the device.

To achieve this, add a new NetworkManager Dispatcher event
'method-change' which is triggered when either an IPv4 or IPv6 method is
changed and set the environment variables IPV4_METHOD and IPV6_METHOD to
the new method so scripts can act on it.