glib-aux: add NMPtrArray helper
Add a simple structure that can own a list of pointers.
The main purpose is to return one or more pointers from a function,
where ownership gets transferred.

This is a more limited form of a GPtrArray. Except, it actually NULL
terminates the list, and it requires only one heap allocation for
the structure and the list.