platform: properly handle `ip route replace` for ignored routes
We don't cache certain routes, based on the protocol. This is a
performance optimization to ignore routes that we usually don't configure.

Still, if the user does `ip route replace` with such a route, then we
need to pass it to nmp_cache_update_netlink_route(), so that we can
properly replace the route. Actually, the implementation will just set
the "resync_required" flag to indicate that a sync is necessary :(

This also means, if we ever write a BPF filter to filter out such routes,
we cannot do that if the route replaces another route.