contrib/rpm: add "22-wifi-mac-addr.conf" to F40+
Install a configuration snippet on Fedora 40+, that sets the default for
"wifi.cloned-mac-address" to "stable" (otherwise, the built-in default
is "preserve").

This will mean, that on Wi-Fi profiles that don't explicitly override
the property "wifi.cloned-mac-address", a stable address is generated.
The benefit is, that Fedora will randomize the MAC address by default.

Note that this also affects all pre-existing Wi-Fi profiles, that don't
explicitly configure something else.

Also set "connection.stable-id=${NETWORK_SSID}" so that the identifier
is generated based on the SSID instead of the connection UUID. As this
is only supported since 1.46, add a [.config] section.

Note that changing the default stable ID also affects other properties:

    ipv6.addr-gen-mode=stable-privacy
    {ethernet,wifi}.cloned-mac-address=stable
    ipv4.dhcp-client-id=stable
    ipv6.dhcp-duid=stable-{llt,ll,uuid}
    {ipv4,ipv6}.iaid=stable

Especially with "ipv6.addr-gen-mode=stable", this change has wider
ramifications.

Depending on how you see it, this is desirable. Randomization should be
done, unless the user opts-out (not the other way around).

This means, after upgrade to F40, different MAC addresses will be used
on most users' Wi-Fi. This means, DHCP might hand out different IP
addresses, sessions might expire, and configuration that depended on the
previous MAC address will be affected.

https://pagure.io/fedora-workstation/issue/350