gitlab-ci: trigger container-prep stage only manually for manual tiers
Previously, the container-prep ran automatically. The benefit is, that the other
stages which depend on the prep, were immediately ready to be started manually.
It only required one click.

The downside was, that we ran those prep stages every time, even if most of
the time the follow-up stages are not started.

Now only trigger the prep stage manually. It is now more cumbersome to
start the later test stages, because we will first have to start the
prep stages, then wait, then click to start the manual stages. But as we
do that less frequently, it seems acceptable.