build: Add fhs3_rundir build option
Distributions trying to comply to FHS (Filesystem Hierarchy Standard),
already use a /var/run -> /run symlink, currently. Despite the move of
the rundir from under /var to the / root directory in the current
version of FHS - 3.0, released in 2015 -, symlinking /var/run to /run is
allowed for backwards compatibility.

Introducing this build option to allow distributions to use /run instead
of /var/run, while keeping /var/run ($localstatedir/run) as the default,
is a good step to ease migration.