docs: allow comma separated list of tags in properties docs
Allow to define a comma separated list of tags, like
"---nmcli,keyfile,dbus---", instead of having to define separate blocks
of comments for each one ("---nmcli---", "---keyfile---"), even for the
same content.

This can be useful to avoid repeating the same comment if 2 targets
needs the same modifications to the general documentation, although this
case might be rare.

It will be more useful to allow using the new "expand enumvals" feature
because it can't be used from gtkdoc comments.