nm-meta-setting-desc: allow specifying reduntant enum type
The get_gtype field in property_typ_data was intended to specify an enum
type for properties that are really defined as (u)int in the NMSetting
class. Specifying get_gtype for properties that are already defined as
enum in the NMSetting class was rejected as an error, detected at
runtime.

However, although it is redundant, it is no harmful to accept a get_gtype
if it's the same type defined in the NMSetting, so let's allow it. Now,
an error will be triggered only if the get_gtype returns a wrong type.
The advantage of doing it this way is that a person adding a new
property to nm-meta-setting-desc might expect that this is allowed, and
since this error was only detected at runtime, there is a small risk of
it ending in a production build.