l3cfg: handle dynamic added routes tracking and deletion
Dynamic added routes i.e ECMP single-hop routes, are not managed by l3cd
as the other ones. Therefore, they need to be tracked properly and
marked as dirty when they are.

This way, NetworkManager makes sure there are no leftovers from merging
ECMP multi-hop routes.

A "os_non_dynamic" flag is required.

A ObjStateData is alive as long as it either has a static route (from
the combined l3cd) or a dynamic (from nm_netns_ip_route_ecmp_commit()).
Note that it can be both!

So, _obj_states_track_prune_dirty() (formerly
_obj_states_remove_track()) needs to be able to look at an obj_state and
determine whether it's still kept alive (by either).

You might think, "but it cannot be both". The routes from the
non-dynamic combined-l3cd are IPv4 single-hop routes with a weight of
zero. On the other hands, the routes from nm_netns_ip_route_ecmp_commit()
are the ones we passed to NMNetns that didn't find an ECMP merge
partner. They should have a positive "weight", and compare different.

Wrong. The route that comes back from nm_netns_ip_route_ecmp_commit()
must have a weight of zero too. The following commit will normalize
the weight of those routes to zero. Then indeed the non-dynamic and
dynamic routes both have weight zero, and they compare equal.

Just try it by configuring two routes that are identical except one has
weight zero and the other weight 42. (note that to see it, you will need
to follow-up patch to normalize the weight).

But why normalize the weight? Because when we get a route back from
nm_netns_ip_route_ecmp_commit(), we want to be able to look into the
platform cache (whether the route exists). NML3Cfg also tracks routes
from platform, and associates them with the obj-state. The Routes from
platform all have "weight" zero. Normalizing the route from
ecmp_commit() is necessary to be the route we want to add and as it
appears in platform (that is, the route with a zero weight).