Allow any kind of VPN as secondaries (Fixed #1395)
As of now, NM allows only connections derived from the
VPN plugin to be used as secondaries. This excludes
non-VPN-plugin connections still providing VPN
functionality, such as Wireguard, to be used as secondary.
This commit extends NetworkManager to allow any kind of VPN
connection to be used as secondary connection.