wwan/ofono: also defer activation if we lose signal while activating
It's possible that while we're waiting for an oFono context to be
activated, the phone becomes disconnected from GPRS. This means the
PropertyChanged signal we're waiting for will never arrive, which will
leave the NM device in prepare state forever.

And since oFono's GPRS attached status can change quite quickly, it's
possible that multiple connection attempts can fail in a row, depleting
autoconnect retries. To avoid that, we treat GPRS un-attach event as a
transient error and defer the context activation again.

Don't stop and restart the timeout source each time we defer the
activation, as multiple defer cycles can elongate the activation time
forever. Instead, put a time bound to everything and use a single
timeout source for such a connection attempt.

For this to work, we must be able to track reliably whether we're in a
middle of a context activation or not. So make sure that whenever we
succeed or give up we clear both the timeout source and the cancellable.