wwan/ofono: correctly get remote error to check for .InProgress
The old code seemed to try to take a shortcut while detecting the error,
but made a mistake in doing so, which mean it was never detected
correctly. Instead, use a helper function provided by Gio itself to do
the job.

And while we're at it, log when we detect such a failure to know exactly
what error we get from oFono.