systemd: add extra capabilities when building with CLAT
Since 5.8, kernel requires CAP_BPF for processes that want to use
eBPF. CAP_PERFMON is also required because the BPF program uses
pointer arithmetic. It would be possible to rewrite the program and
the parsing helpers to avoid pointer arithmetic, but that requires
some effort.

Add the capabilities to the service unit when we are building with CLAT
support.