libnm,nmcli: introduce a ipv4.dhcp-use-routes property
The DHCP server can provide routes to clients using different options:

 - 3 (Router): only for the default gateway
 - 121 (Classless Static Route)
 - 249 (Microsoft Classless Static Route)
 - 33 (Static Route)

TunnelVision [1] is a recently published technique that allows a
malicious actor to reroute the victim's traffic that should normally
directed to a VPN, by injecting higher priority routes into the
victim's routing table via the options mention above.

A defense to this attack is to ignore the non-default routes from
those options; introduce a ipv4.dhcp-use-routes property for this
purpose.

The property is currently implemented on the NMSettingIpConfig
setting, and so it can be potentially implemented also for IPv6 in the
future. At the moment it's not because DHCPv6 doesn't provide routes.

[1]  https://github.com/leviathansecurity/TunnelVision/

https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/issues/1535