device: introduce ipv4.link-local=until_dhcp
This is proof of concept code to set an IPv4 link-local address until
a DHCP lease has been obtained and drop it at that point.

The current ipv4.link-local=enabled setting probably violates RFC3927[1]
which says:
  a host SHOULD NOT have both an operable routable address and an IPv4
  Link-Local address configured on the same interface.

This implements a solution that is more standard and conforms to what others are doing.

Fixes: #1562
Link: https://github.com/systemd/systemd/issues/13316
Link: https://datatracker.ietf.org/doc/html/rfc3927#section-1.9 [1]