manager: don't renew dhcp lease when software devices' MAC is empty
If a bridge device doesn't have a permanent MAC address, its MAC address
will change based on the attached ports. When suspending the system, all
slaves can be released, thus MAC address of the bridge interface changes
to 00:00:00:00:00:00. In such scenarios, if you resume the system,
NetworkManager tries to renew dhcp lease on the bridge interface with an
empty MAC address. Because its slaved ports haven't been attached yet.

Fix that in this commit.

https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/issues/1587