build: make manpages a mandatory part of the build by default
Currently, both man pages and gtk-doc HTML documentation are enabled by
the same build option. It is common for users to want to choose whether
to build HTML docs, as not everyone cares about HTML developer docs, but
manpages are intended directly for end-user consumption and should
always be available.

Prebuilt manpages are available and installed as part of release dist
tarballs. When building from a git checkout, manpages require enabling
introspection. There is never a good reason *not* to install the former.
In the latter case, require users to specifically opt out of manpages
via a configure option.

Resolves: https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/issues/1653