udev: add rules file for BMC USB CDC network interfaces
Some systems contain a USB link between the BMC and host which the BMC
configures as a CDC network connection. Unfortunately each time the BMC
boots it generates a new link which the host doesn't associate with any
previous connection.

To avoid having NetworkManager repeatedly try and fail to get an address
via DHCP, add a new udev rules file which tells it to configure it as
link-local only.