wifi: handle password change of WPA3 access point
When a WPA2 access point changes PSK, wpa_supplicant tries to reassociate and
the device state goes from 4WAY_HANDSHAKE to DISCONNECTED which triggers to
prompt the user for a the new password. In case of a WPA3 access point, the user
is not prompted for a new password because there is no 4-way handshake, we go
from state AUTHENTICATING to DISCONNECTED. Here we make sure the end-user sees
the same password prompt whether it's WPA2 or WPA3.