Support referring parent by connection.id
Currently, `connection.controller` support referring to connection by
`connection.id`.

This patch extend `vlan.parent` to support the same. With this, it is
possible to activate a VLAN with its parent connection chose by
PCI-address matching.

When multiple connection, NM choose one randomly, so user have to deal
with overlapping on `connection.id`.

Resolves: https://issues.redhat.com/browse/RHEL-94783

Signed-off-by: Gris Ge <fge@redhat.com>